Refactoring after Lournet's comments

+++ src/libaktualizr/storage/sqlstorage.cc
@@ -1048,11 +1048,11 @@ void SQLStorage::clearInstallationResult() {
   }
 }
 
-boost::optional<std::pair<int64_t, size_t>> SQLStorage::checkTargetFile(const Uptane::Target& target) const {
+boost::optional<size_t> SQLStorage::checkTargetFile(const Uptane::Target& target) const {
   SQLite3Guard db = dbConnection();
 
   auto statement = db.prepareStatement<std::string>(
-      "SELECT real_size, sha256, sha512, image_id FROM target_images WHERE filename = ?;", target.filename());
+      "SELECT real_size, sha256, sha512 FROM target_images WHERE filename = ?;", target.filename());
 
   int statement_state;
   while ((statement_state = statement.step()) == SQLITE_ROW) {
@@ -1077,9 +1077,7 @@ boost::optional<std::pair<int64_t, size_t>> SQLStorage::checkTargetFile(const Up
       }
     }
     if (((*sha256).empty() || sha256_match) && ((*sha512).empty() || sha512_match)) {
-      std::pair<int64_t, size_t> info(statement.get_result_col_int(3),
-                                      static_cast<size_t>(statement.get_result_col_int(0)));
-      return boost::optional<std::pair<int64_t, size_t>>(info);
+      return {static_cast<size_t>(statement.get_result_col_int(0))};
     }
   }
 
@@ -1111,13 +1109,12 @@ class SQLTargetWHandle : public StorageTargetWHandle {
         sha512Hash = hash.HashString();
       }
     }
-
-    auto statement =
-        db_.prepareStatement<std::string>("SELECT image_id FROM target_images WHERE filename = ?;", target_.filename());
+    auto statement = db_.prepareStatement<std::string>("SELECT rowid FROM target_images_data WHERE filename = ?;",
+                                                       target_.filename());
 
     if (statement.step() == SQLITE_ROW) {
-      auto image_id = statement.get_result_col_int(0);
-      statement = db_.prepareStatement<int64_t>("DELETE FROM target_images_data WHERE rowid=?;", image_id);
+      statement =
+          db_.prepareStatement<std::string>("DELETE FROM target_images_data WHERE filename=?;", target_.filename());
 
       if (statement.step() != SQLITE_DONE) {
         LOG_ERROR << "Statement step failure: " << db_.errmsg();
@@ -1125,8 +1122,9 @@ class SQLTargetWHandle : public StorageTargetWHandle {
       }
     }
 
-    statement = db_.prepareStatement<SQLZeroBlob>("INSERT OR REPLACE INTO target_images_data (image_data) VALUES (?);",
-                                                  SQLZeroBlob{target_.length()});
+    statement = db_.prepareStatement<std::string, SQLZeroBlob>(
+        "INSERT OR REPLACE INTO target_images_data (filename, image_data) VALUES (?,?);", target_.filename(),
+        SQLZeroBlob{target_.length()});
 
     if (statement.step() != SQLITE_DONE) {
       LOG_ERROR << "Statement step failure: " << db_.errmsg();
@@ -1134,9 +1132,9 @@ class SQLTargetWHandle : public StorageTargetWHandle {
     }
 
     row_id_ = sqlite3_last_insert_rowid(db_.get());
-    statement = db_.prepareStatement<std::string, std::string, std::string, int64_t>(
-        "INSERT OR REPLACE INTO target_images (filename, sha256, sha512, image_id) VALUES (?, ?, ?, ?);",
-        target_.filename(), sha256Hash, sha512Hash, row_id_);
+    statement = db_.prepareStatement<std::string, std::string, std::string>(
+        "INSERT OR REPLACE INTO target_images (filename, sha256, sha512) VALUES ( ?, ?, ?);", target_.filename(),
+        sha256Hash, sha512Hash);
 
     if (statement.step() != SQLITE_DONE) {
       LOG_ERROR << "Statement step failure: " << db_.errmsg();
@@ -1188,7 +1186,6 @@ class SQLTargetWHandle : public StorageTargetWHandle {
         throw StorageTargetWHandle::WriteError("could not update size of " + target_.filename() + " in sql storage");
       }
     }
-
     closed_ = true;
   }
 
@@ -1240,15 +1237,22 @@ class SQLTargetRHandle : public StorageTargetRHandle {
         read_size_(0),
         closed_(false),
         blob_(nullptr) {
-    StorageTargetRHandle::ReadError exc("could not read file " + filename_ + " from sql storage");
+    StorageTargetRHandle::ReadError exc("could not read file " + target_.filename() + " from sql storage");
 
     auto exists = storage.checkTargetFile(target_);
     if (!exists) {
       throw exc;
     }
+    auto statement = db_.prepareStatement<std::string>("SELECT rowid FROM target_images_data WHERE filename = ?;",
+                                                       target_.filename());
 
-    row_id_ = exists->first;
-    size_ = exists->second;
+    if (statement.step() != SQLITE_ROW) {
+      LOG_ERROR << "Statement step failure: " << db_.errmsg();
+      throw std::runtime_error("Could not find target file");
+    }
+
+    row_id_ = statement.get_result_col_int(0);
+    size_ = *exists;
     partial_ = size_ < target_.length();
     if (sqlite3_blob_open(db_.get(), "main", "target_images_data", "image_data", row_id_, 0, &blob_) != SQLITE_OK) {
       LOG_ERROR << "Could not open blob: " << db_.errmsg();
@@ -1295,7 +1299,6 @@ class SQLTargetRHandle : public StorageTargetRHandle {
   boost::filesystem::path db_path_;
   SQLite3Guard db_;
   Uptane::Target target_;
-  const std::string filename_;
   size_t size_;
   size_t read_size_;
   bool closed_;
@@ -1311,13 +1314,17 @@ std::unique_ptr<StorageTargetRHandle> SQLStorage::openTargetFile(const Uptane::T
 void SQLStorage::removeTargetFile(const std::string& filename) {
   SQLite3Guard db = dbConnection();
 
-  auto statement = db.prepareStatement<std::string>("SELECT image_id FROM target_images WHERE filename = ?;", filename);
+  if (!db.beginTransaction()) {
+    LOG_ERROR << "Can't start transaction: " << db.errmsg();
+    return;
+  }
+
+  auto statement = db.prepareStatement<std::string>("SELECT filename FROM target_images WHERE filename = ?;", filename);
 
   if (statement.step() != SQLITE_ROW) {
     LOG_ERROR << "Statement step failure: " << db.errmsg();
     throw std::runtime_error("Could not find target file");
   }
-  auto image_id = statement.get_result_col_int(0);
 
   statement = db.prepareStatement<std::string>("DELETE FROM target_images WHERE filename=?;", filename);
 
@@ -1326,7 +1333,7 @@ void SQLStorage::removeTargetFile(const std::string& filename) {
     throw std::runtime_error("Could not remove target file");
   }
 
-  statement = db.prepareStatement<int64_t>("DELETE FROM target_images_data WHERE rowid=?;", image_id);
+  statement = db.prepareStatement<std::string>("DELETE FROM target_images_data WHERE filename=?;", filename);
 
   if (statement.step() != SQLITE_DONE) {
     LOG_ERROR << "Statement step failure: " << db.errmsg();
@@ -1336,6 +1343,8 @@ void SQLStorage::removeTargetFile(const std::string& filename) {
   if (sqlite3_changes(db.get()) != 1) {
     throw std::runtime_error("Target file " + filename + " not found");
   }
+
+  db.commitTransaction();
 }
 
 void SQLStorage::cleanUp() { boost::filesystem::remove_all(dbPath()); }
